feat: remove from component to validation file
